refactor: New DeviceNode classes replacing former Device class, new Device classes will be used for intrinsic operation logic of the devices
