+++ src/graphics/left_bar.ts
@@ -3,8 +3,8 @@ import ComputerSvg from "../assets/pc.svg";
 import SwitchSvg from "../assets/switch.svg";
 import { addDevice } from "../types/viewportManager";
 import { GlobalContext } from "../context";
-import { DeviceType } from "../types/devices/device";
-import { Layer, layerFromName } from "../types/devices/layer";
+import { DeviceType } from "../types/deviceNodes/deviceNode";
+import { Layer, layerFromName } from "../types/layer";
 
 export class LeftBar {
   private leftBar: HTMLElement;
