+++ src/index.ts
@@ -27,7 +27,7 @@ import PlaySvg from "./assets/play-icon.svg";
 import PauseSvg from "./assets/pause-icon.svg";
 import UndoSvg from "./assets/left-curve-arrow.svg";
 import RedoSvg from "./assets/right-curve-arrow.svg";
-import { layerToName } from "./types/devices/layer";
+import { layerToName } from "./types/layer";
 
 const assets = [
   RouterSvg,
