+++ src/types/devices/switch.ts
@@ -1,48 +1,10 @@
-import { Device, DeviceType, Layer } from "./device";
-import { ViewGraph } from "../graphs/viewgraph";
-import SwitchImage from "../../assets/switch.svg";
-import { Position } from "../common";
-import { DeviceInfo, RightBar } from "../../graphics/right_bar";
+import { IPv4Packet } from "../../packets/ip";
 import { DeviceId } from "../graphs/datagraph";
 import { Packet } from "../packet";
-import { Texture } from "pixi.js";
-import { MacAddress } from "../../packets/ethernet";
-import { IPv4Packet } from "../../packets/ip";
-import { GlobalContext } from "../../context";
+import { Device } from "./device";
 
 export class Switch extends Device {
-  static DEVICE_TEXTURE: Texture;
-
-  static getTexture() {
-    if (!Switch.DEVICE_TEXTURE) {
-      Switch.DEVICE_TEXTURE = Texture.from(SwitchImage);
-    }
-    return Switch.DEVICE_TEXTURE;
-  }
-
-  constructor(
-    id: DeviceId,
-    viewgraph: ViewGraph,
-    ctx: GlobalContext,
-    position: Position,
-    mac: MacAddress,
-  ) {
-    super(id, Switch.getTexture(), viewgraph, ctx, position, mac);
-  }
-
-  showInfo(): void {
-    const info = new DeviceInfo(this);
-    info.addEmptySpace();
-    RightBar.getInstance().renderInfo(info);
-  }
-
-  getLayer(): Layer {
-    return Layer.Link;
-  }
-
-  getType(): DeviceType {
-    return DeviceType.Switch;
-  }
+  viewgraph: any;
 
   receivePacket(packet: Packet): Promise<DeviceId | null> {
     const datagram = packet.rawPacket.payload;
