+++ src/types/speedMultiplier.ts
